refactor: Remove Intelligence tab and add Anweisungen section
- Remove deprecated Mem0ry/Intelligence tab and associated styles
- Add new AnweisungenSection component for content management
- Clean up profile page navigation and constants
- Update DocumentOverview and useWolkeSync hooks
