--- gruenerator_frontend/src/assets/styles/features/auth/mem0ry-tab.css
@@ -1,283 +1,0 @@
-/* Mem0ry Tab Specific Styles */
-
-.group-description {
-  margin: var(--spacing-small) 0;
-  color: var(--font-color-disabled);
-  font-size: 0.95rem;
-  line-height: 1.4;
-}
-
-.empty-state-icon {
-  color: var(--interactive-accent-color);
-  margin-bottom: var(--spacing-medium);
-}
-
-.empty-state-description {
-  margin-top: var(--spacing-small);
-  color: var(--font-color-disabled);
-  font-size: 0.9rem;
-  line-height: 1.4;
-}
-
-.knowledge-empty-state {
-  text-align: center;
-  padding: var(--spacing-large);
-  background-color: var(--background-color-alt);
-  border-radius: var(--card-border-radius-medium);
-  margin-top: var(--spacing-medium);
-}
-
-.memory-content {
-  padding: var(--spacing-medium);
-  background-color: var(--card-background);
-  border-radius: var(--card-border-radius-small);
-  border: var(--card-border);
-  transition: var(--input-transition);
-}
-
-
-.memory-text {
-  font-weight: 500;
-  margin-bottom: var(--spacing-small);
-  color: var(--font-color);
-  line-height: 1.5;
-  margin: 0;
-}
-
-.memory-metadata {
-  font-size: 0.85rem;
-  color: var(--font-color-disabled);
-  margin-top: var(--spacing-small);
-  padding-top: var(--spacing-small);
-  border-top: var(--border-subtle);
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-small);
-  flex-wrap: wrap;
-}
-
-.metadata-badge {
-  display: inline-block;
-  padding: 2px 8px;
-  background-color: var(--hover-color-alt);
-  color: var(--interactive-accent-color);
-  border-radius: var(--card-border-radius-small);
-  font-size: 0.75rem;
-  font-weight: 500;
-  text-transform: capitalize;
-}
-
-.memory-timestamp {
-  font-size: 0.8rem;
-  color: var(--font-color-disabled);
-  opacity: 0.8;
-}
-
-/* Profile actions with icons */
-.profile-actions button {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-xsmall);
-}
-
-.profile-actions svg {
-  flex-shrink: 0;
-}
-
-/* Auth error message with icons */
-.auth-error-message {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-xsmall);
-  margin: var(--spacing-medium) 0;
-}
-
-.auth-error-message svg {
-  flex-shrink: 0;
-}
-
-/* Profile content centered for loading */
-.profile-content-centered {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: var(--spacing-large);
-  min-height: 200px;
-}
-
-/* Enhanced memory list styling */
-.memories-list .knowledge-entry {
-  transition: all 0.25s ease-out;
-}
-
-.memories-list .knowledge-entry:hover {
-  transform: translateY(-1px);
-}
-
-.memories-list .form-field-wrapper {
-  margin-bottom: 0;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-  .memory-metadata {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: var(--spacing-xsmall);
-  }
-  
-  .metadata-badge {
-    align-self: flex-start;
-  }
-  
-  .profile-actions {
-    flex-direction: column;
-  }
-  
-  .profile-actions button {
-    width: 100%;
-    justify-content: center;
-  }
-}
-
-
-
-/* Memory Header with Delete Button */
-.memory-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  gap: var(--spacing-small);
-}
-
-.delete-memory-button {
-  background: none;
-  border: none;
-  color: var(--error-red);
-  cursor: pointer;
-  padding: var(--spacing-xsmall);
-  border-radius: var(--card-border-radius-small);
-  transition: var(--input-transition);
-  flex-shrink: 0;
-  opacity: 0.7;
-}
-
-.delete-memory-button:hover {
-  opacity: 1;
-  background-color: var(--input-background);
-  transform: scale(1.05);
-}
-
-.delete-memory-button:focus {
-  outline: 2px solid var(--error-red);
-  outline-offset: 2px;
-}
-
-/* Title Actions Container */
-.title-actions {
-  display: flex;
-  gap: var(--spacing-small);
-  align-items: center;
-}
-
-/* Form Group Title Container - for header with button on right */
-.form-group-title-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.add-knowledge-button.danger {
-  background-color: var(--error-red);
-  color: var(--white);
-}
-
-.add-knowledge-button.danger:hover {
-  background-color: var(--grey-800);
-}
-
-/* Form Help Text */
-.form-help-text {
-  margin-top: var(--spacing-xsmall);
-  font-size: 0.9rem;
-  color: var(--font-color-disabled);
-  line-height: 1.4;
-}
-
-/* Memory Settings Toggle Spacing */
-.form-group .feature-toggle {
-  margin-top: var(--spacing-large);
-  margin-bottom: var(--spacing-medium);
-}
-
-/* Delete All Memories Link */
-.delete-all-container {
-  margin-top: var(--spacing-medium);
-  text-align: center;
-  padding-top: var(--spacing-medium);
-  border-top: var(--border-subtle);
-}
-
-.delete-all-link {
-  background: none;
-  border: none;
-  color: var(--error-red);
-  cursor: pointer;
-  font-size: 0.85rem;
-  text-decoration: underline;
-  padding: var(--spacing-xsmall);
-  border-radius: var(--card-border-radius-small);
-  transition: var(--input-transition);
-  opacity: 0.8;
-}
-
-.delete-all-link:hover {
-  opacity: 1;
-  background-color: var(--input-background);
-  text-decoration: none;
-}
-
-.delete-all-link:focus {
-  outline: 2px solid var(--error-red);
-  outline-offset: 2px;
-}
-
-/* Accessibility improvements */
-@media (prefers-reduced-motion: reduce) {
-  .memory-content,
-  .memories-list .knowledge-entry,
-  .delete-memory-button,
-  .delete-all-link {
-    transition: none;
-  }
-  
-  .memory-content:hover,
-  .memories-list .knowledge-entry:hover,
-  .delete-memory-button:hover {
-    transform: none;
-  }
-}
-
-/* Mobile Responsive */
-@media (max-width: 768px) {
-  .memory-header {
-    flex-direction: column;
-    align-items: stretch;
-  }
-  
-  .delete-memory-button {
-    align-self: flex-end;
-    margin-top: var(--spacing-xsmall);
-  }
-  
-  .title-actions {
-    flex-direction: column;
-    width: 100%;
-  }
-  
-  .title-actions button {
-    width: 100%;
-    justify-content: center;
-  }
-}
