+++ gruenerator_frontend/src/components/utils/documentOverviewUtils.js
@@ -17,8 +17,20 @@ export const truncateForPreview = (content, maxLength = 300) => {
   }
 };
 
+// Cache for memoized markdown stripping - prevents expensive regex operations on re-renders
+const stripMarkdownCache = new Map();
+const CACHE_MAX_SIZE = 500;
+
 export const stripMarkdownForPreview = (content, maxLength = 300) => {
   if (!content || typeof content !== 'string') return '';
+
+  // Create cache key from content hash (first 100 chars + length for uniqueness)
+  const cacheKey = `${content.slice(0, 100)}_${content.length}_${maxLength}`;
+
+  // Return cached result if available
+  if (stripMarkdownCache.has(cacheKey)) {
+    return stripMarkdownCache.get(cacheKey);
+  }
 
   let cleaned = content
     .replace(/^#{1,6}\s+/gm, '')              // # Headers
@@ -35,7 +47,17 @@ export const stripMarkdownForPreview = (content, maxLength = 300) => {
     .replace(/~~([^~]+)~~/g, '$1')            // ~~strikethrough~~
     .trim();
 
-  return truncateForPreview(cleaned, maxLength);
+  const result = truncateForPreview(cleaned, maxLength);
+
+  // Cache the result (with size limit to prevent memory leaks)
+  if (stripMarkdownCache.size >= CACHE_MAX_SIZE) {
+    // Clear oldest entries (simple strategy: clear first 100)
+    const keys = Array.from(stripMarkdownCache.keys()).slice(0, 100);
+    keys.forEach(key => stripMarkdownCache.delete(key));
+  }
+  stripMarkdownCache.set(cacheKey, result);
+
+  return result;
 };
 
 export const formatDate = (dateString) => {
